Revert "Check and store parameters in aleo directory"
