+++ parameters/Cargo.toml
@@ -27,9 +27,6 @@ path = "../utilities"
 version = "0.7.5"
 default-features = false
 
-[dependencies.aleo-std]
-version = "0.1.0"
-
 [dependencies.anyhow]
 version = "1"
