+++ parameters/src/lib.rs
@@ -26,8 +26,6 @@ extern crate thiserror;
 #[macro_use]
 pub mod macros;
 
-pub mod check_parameters;
-
 pub mod errors;
 pub use errors::*;
